requirements.txt
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.155.0
anthropic>=0.40.0
pydantic-settings>=2.7.0
